python/requirement.txt
evo
torchmetrics
numpy
scipy
scikit-image
lpips
pillow
tqdm
plyfile
